GroupController - finish endpoints
