+++ src/main/java/com/kodilla/ecommercee/dto/GroupDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.dto;
+
+public class GroupDto {
+}
